Revert "Transient storage for sensitive fields"
This reverts commit 0936e93fc5d67634251b1561698e17e605755c8d.
