+++ opba-protocols/xs2a-protocol/src/main/java/de/adorsys/opba/protocol/xs2a/config/flowable/CanSerializeUtil.java
@@ -1,0 +1,14 @@
+package de.adorsys.opba.protocol.xs2a.config.flowable;
+
+import lombok.experimental.UtilityClass;
+
+import java.util.List;
+
+@UtilityClass
+@SuppressWarnings("checkstyle:HideUtilityClassConstructor") // Lombok generates private ctor.
+public class CanSerializeUtil {
+
+    public boolean canSerialize(String canonicalName, List<String> allowOnlyClassesWithPrefix) {
+        return allowOnlyClassesWithPrefix.stream().anyMatch(canonicalName::startsWith);
+    }
+}
