+++ opba-protocols/xs2a-protocol/src/main/java/de/adorsys/opba/protocol/xs2a/config/flowable/Xs2aFlowableProperties.java
@@ -21,6 +21,6 @@ public class Xs2aFlowableProperties {
     private int maxLength = 2048;
 
     public boolean canSerialize(String canonicalName) {
-        return SerializerUtil.canSerialize(canonicalName, serializeOnlyPackages);
+        return CanSerializeUtil.canSerialize(canonicalName, serializeOnlyPackages);
     }
 }
