--- opba-protocols/xs2a-protocol/src/main/java/de/adorsys/opba/protocol/xs2a/service/storage/NeedsTransientStorage.java
@@ -1,6 +1,0 @@
-package de.adorsys.opba.protocol.xs2a.service.storage;
-
-public interface NeedsTransientStorage {
-
-    void setTransientStorage(TransientDataStorage storage);
-}
