+++ opba-protocols/xs2a-protocol/src/main/java/de/adorsys/opba/protocol/xs2a/service/xs2a/context/Xs2aContext.java
@@ -1,9 +1,6 @@
 package de.adorsys.opba.protocol.xs2a.service.xs2a.context;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import de.adorsys.opba.protocol.xs2a.domain.dto.forms.ScaMethod;
-import de.adorsys.opba.protocol.xs2a.service.storage.NeedsTransientStorage;
-import de.adorsys.opba.protocol.xs2a.service.storage.TransientDataStorage;
 import de.adorsys.xs2a.adapter.service.model.AuthenticationObject;
 import de.adorsys.xs2a.adapter.service.model.StartScaProcessResponse;
 import lombok.Data;
@@ -14,10 +11,7 @@ import java.util.List;
 // TODO - Make immutable, modify only with toBuilder
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class Xs2aContext extends BaseContext implements NeedsTransientStorage {
-
-    @JsonIgnore
-    private TransientDataStorage transientStorage;
+public class Xs2aContext extends BaseContext {
 
     // Mandatory static
     private String psuId;
@@ -35,6 +29,10 @@ public class Xs2aContext extends BaseContext implements NeedsTransientStorage {
     private List<ScaMethod> availableSca;
     private String userSelectScaId;
     private AuthenticationObject scaSelected;
+    private String lastScaChallenge;
+
+    // sensitive - do not persist?
+    private String psuPassword;
 
     // TODO: protect from overriding using reflection https://github.com/adorsys/open-banking-gateway/issues/251
     private String redirectUriOk;
@@ -46,22 +44,4 @@ public class Xs2aContext extends BaseContext implements NeedsTransientStorage {
     private String fintechRedirectUriNok;
 
     private boolean redirectConsentOk;
-
-    public String getPsuPassword() {
-        TransientDataStorage.DataEntry entry = transientStorage.get(this);
-        return null != entry ? transientStorage.get(this).getPsuPassword() : null;
-    }
-
-    public String getLastScaChallenge() {
-        TransientDataStorage.DataEntry entry = transientStorage.get(this);
-        return null != entry ? transientStorage.get(this).getScaChallengeResult() : null;
-    }
-
-    public void setPsuPassword(String psuPassword) {
-        transientStorage.set(this, new TransientDataStorage.DataEntry(psuPassword, null));
-    }
-
-    public void setLastScaChallenge(String scaChallengeResult) {
-        transientStorage.set(this, new TransientDataStorage.DataEntry(null, scaChallengeResult));
-    }
 }
