+++ opba-protocols/xs2a-protocol/src/main/resources/application.yml
@@ -38,8 +38,6 @@ protocol:
       select-sca-method: 'http://localhost:5500/parameters/select-sca-method/#{execution.getId()}?q=#{urlSafe(context.getAvailableSca().toString())}'
       report-sca-result: 'http://localhost:5500/parameters/report-sca-result/#{execution.getId()}?q=#{urlSafe(context.getScaSelected().getAuthenticationType() + ":" + context.getScaSelected().getName())}'
   gateway-base-url: http://localhost:8085
-  expirable:
-    expire-after-write: PT600S
 
 pkcs12:
   keyStore: sample-qwac.keystore
